GraphTransformer

Documentation for GraphTransformer.

GraphTransformer.FFN

GraphTransformer.GraphTransformerLayer

GraphTransformer.message

GraphTransformer.FFN — Type

FeedForwardNetwork

GraphTransformer.GraphTransformerLayer — Type

GraphTransformerLayer

GraphTransformer.message — Method

Arguments:

x_i: Feature vector of i'th node

x_j: Feature vector of j'th node, where j is in the neighbourhood of i

e_ij: Feature vector for the edge between i and j